+++ src/Styles/Sass/App.sass
@@ -6,7 +6,7 @@
 
 body
     background: #71a5de
-    font-size: 14px
+    font-size: 10px
     padding: .2rem
     @include columnFlex
     justify-content: center
@@ -40,6 +40,7 @@ nav
     padding: .2rem
     gap: .2rem
     justify-content: center
+
     align-self: stretch
 
 .navBar
@@ -47,8 +48,12 @@ nav
     justify-content: space-between
     align-items: center
     border: 2px solid $secondarycolor
-    padding: .2rem .5rem
+    padding: .25rem
     align-self: stretch
+
+    h2
+        font-weight: 700 !important
+        font-size: 1em
 
     div
         @include rowFlex
@@ -62,10 +67,22 @@ nav
 
 .mainSect
     @include columnFlex
-    justify-content: stretch
+    justify-content: center
     align-items: center
     gap: 2rem
-
+.HomeComponent
+    text-align: left
+    white-space: normal !important
+    margin: 2rem 3rem
+    h2
+        font-size: 1.2rem
+        font-weight: 600
+    p
+        font-size: 1rem
+        margin-left: .5rem
+    ol li
+        font-size: 1rem
+        margin-left: 3rem
 .ContactComponent
     @include columnFlex
     justify-content: center
