+++ src/components/Home.js
@@ -1,0 +1,20 @@
+const Home = () => {
+  const result = (
+    <div className="HomeComponent">
+      <h2>Geeky Calculator</h2>
+      <p>
+        Welcome to my first single page application(SPA) using react framework.
+        this app has two main sections:
+      </p>
+      <ol>
+        <li> The Calculator App </li>
+        <li> Quotes Section which shows you random quotes about success</li>
+      </ol>
+      <p>I hope you like it and find it useful.</p>
+    </div>
+  );
+
+  return result;
+};
+
+export default Home;
